scripts/used-images/requirements.txt
gitpython
tqdm
boto3
